C-JIT demo [cross-]compiling a JIT shared library.
... using gcc/clang to JIT-compile for host,
or ncc/clang -target ve-linux... to cross-compile.

API is to pass a std::string ccode to a DllPipe, which
creates a lib<how>_basename.so shared library.
Then dlopen/dlsym to execute to JIT function.
'make dltest1.log' to run the demo.
For all variants of dltest1 to run, you may need:
gcc,g++,ncc,nc++,clang and clang for VE.